requirements.txt
flask==3.1.3
flask-cors==6.0.2
flask-sqlalchemy==3.1.1
flask-limiter==4.1.1
requests==2.33.1
beautifulsoup4==4.14.3
openai==2.31.0
python-dotenv==1.2.2
gunicorn==25.3.0
cloudscraper==1.2.71
pdfplumber==0.11.8
PyPDF2==3.0.1
apscheduler==3.11.0
